Populating the host_name field with cluster-local-ips.
Summary:
In the current implementation, the hostname is set to empty-string in the k8s scheduler's describe function.  This diff uses the k8s client to map the pod-name to its ip, and generate the pop cluster-dns <dashed-ip>.<namespace>.cluster.local and return it.  If it fails, it makes a best attempt at a fallback. 

Note: The fallback does not match prior functionality; but diff #4 in this stack makes it fall back to empty-string and adds tests to cover this diff.

Differential Revision: D88567449