Adding opts to torchx k8s scheduler to allow overriding cpu/memory overhead and AWS EFA device count.
Summary:
This diff adds support for new scheduler ops {reserved_millicpu, reserved_memmb, efa_device_count}.  This lets you override the overhead CPU/memory/efa-devices for established resources.

The first two are important as in kubernetes, the amount of schedulable room on a node is determined by its size *and* the other things running on it.  Different clusters will have different security, logging, etc tools meaning that they will have different practically schedulable space per node.  If torchx assumes only 100m (1/10th) of a node is required for headroom, clusters with more daemon sets for security/logging/etc (i.e. most enterprise clusters) will not be able to schedule its pods.

For EFA specifically, clusters need to be built to support it and it can be off or on.  So, it is good to assume a cluster uses it if a node supports it, but we should be able to fix it.

Differential Revision: D88564180