Populating the host_name field with cluster-local-ips. (#1173)
Summary:

In the current implementation, the hostname is set to empty-string in the k8s scheduler's describe function (basically it's an implementation gap).  This diff uses the k8s client to map the pod-name to its ip, and generate the pop cluster-dns <dashed-ip>.<namespace>.cluster.local and return it.

If the pod IP is not available yet, this version of the diff will simply return an empty hostname since the k8s hostname is derived from the IP.  The describe API is intended to quickly return the actual state from the scheduler's POV; so returning an empty host name when the IP is not yet assigned is better than blocking until it is, which could take an arbitrary amount of time.

Existing users of the API should be fine as the hostname was not populuated anyway.  New users will need to implement polling until the host name is ready, if they have not already.

Note: adding extra test and resubmitting to avoid GitHub coverage block.

Reviewed By: AbishekS

Differential Revision: D88567449