Feature to add cluster-auth as an option (i.e. k8s service account support). (#1172)
Summary:

This diff adds cluster-auth as an option in the k8s scheduler by first checking if it works, and then falling back to the old auth scheme.  So, if we are in pod in a k8s cluster and the pod has a k8s service-account, it will use that.  If it does not, it will try to fallback to a kubeconfig.

Reviewed By: AbishekS

Differential Revision: D88568003